# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: EVC Mesh API
  description: |
    REST API for EVC Mesh -- a multi-agent task management platform.
    Provides workspace, project, task, comment, artifact, agent, and event-bus
    endpoints for both human users (JWT) and AI agents (API key).
  version: "0.1.0"
  contact:
    name: EVC Mesh Team
servers:
  - url: http://localhost:8005
    description: Local development
security:
  - BearerAuth: []
  - AgentKeyAuth: []
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained via /auth/login or /auth/refresh.
    AgentKeyAuth:
      type: apiKey
      in: header
      name: X-Agent-Key
      description: Agent API key in the format agk_{workspace_slug}_{random}.

  # ------------------------------------------------------------------
  # Reusable schemas
  # ------------------------------------------------------------------
  schemas:
...
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties:
            type: string

    # ---------- Pagination wrapper ----------
...
    Rule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
          nullable: true
        scope:
          type: string
          enum: [workspace, project]
        rule_type:
          type: string
        name:
          type: string
        description:
          type: string
        config:
          type: object
        applies_to_actor_types:
          type: array
          items:
            type: string
        applies_to_roles:
          type: array
          items:
            type: string
        enforcement:
          type: string
          enum: [advisory, block]
        priority:
          type: integer
        is_enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /api/v1/rules/{rule_id}:
    parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [GovernanceRules]
      summary: Get a governance rule by ID
      operationId: getRule
      responses:
        "200":
          description: Rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "404":
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags: [GovernanceRules]
      summary: Update a governance rule
      operationId: updateRule
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                config:
                  type: object
                applies_to_actor_types:
                  type: array
                  items:
                    type: string
                applies_to_roles:
                  type: array
                  items:
                    type: string
                enforcement:
                  type: string
                  enum: [advisory, block]
                priority:
                  type: integer
                is_enabled:
                  type: boolean
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "404":
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [GovernanceRules]
      summary: Delete a governance rule
      operationId: deleteRule
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Rule deleted
        "404":
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...